+++ src/lib/djerba/core/constants.py
@@ -7,6 +7,9 @@ DJERBA_TEST_DIR_VAR = 'DJERBA_TEST_DIR'
 
 # shared constants for core classes
 ATTRIBUTES = 'attributes'
+DEPENDS_CONFIGURE = 'depends_configure'
+DEPENDS_EXTRACT = 'depends_extract'
+# render dependencies are intentionally not defined
 CLINICAL = 'clinical'
 SUPPLEMENTARY = 'supplementary'
 CONFIGURE_PRIORITY = 'configure_priority'
@@ -17,7 +20,14 @@ PRIORITY_KEYS = [
     EXTRACT_PRIORITY,
     RENDER_PRIORITY
 ]
-
+RESERVED_PARAMS = [
+    ATTRIBUTES,
+    DEPENDS_CONFIGURE,
+    DEPENDS_EXTRACT,
+    CONFIGURE_PRIORITY,
+    EXTRACT_PRIORITY,
+    RENDER_PRIORITY
+]
 PRIORITIES = 'priorities'
 CONFIGURE = 'configure'
 EXTRACT = 'extract'